+++ zioServer/src/main/scala/tester/srv/dao/CourseDao.scala
@@ -16,23 +16,23 @@ import java.time.Instant
 
 object CourseDao extends AbstractDao[Course]
   with ById[Course] {
-  case class Course(id: Long, userId: Long, templateAlias: String, seed: Long, startedAt: Option[Instant], endedAt: Option[Instant])
+  case class Course(id: Int, userId: Int, templateAlias: String, seed: Int, startedAt: Option[Instant], endedAt: Option[Instant])
 
   override val schema: Schema[Course] = DeriveSchema.gen[Course]
   override val tableName: String = "Course"
 
-  def byAliasAndUserId(alias: String, userId: Long): TranzactIO[Option[Course]] = 
+  def byAliasAndUserId(alias: String, userId: Int): TranzactIO[Option[Course]] =
     selectWhereAndOption(fr"userId = $userId", fr"templateAlias = $alias")
 
-  def activeUserCourses(userId: Long): TranzactIO[List[Course]] = 
+  def activeUserCourses(userId: Int): TranzactIO[List[Course]] =
     selectWhereList(fr"""userID = $userId AND (startedAt = NULL OR startedAt < NOW()::TIMESTAMP) AND
          (endedAt = NULL OR NOW()::TIMESTAMP < endedAt)""")
   
 
-  def previousUserCourses(userId: Long): TranzactIO[List[Course]] =
+  def previousUserCourses(userId: Int): TranzactIO[List[Course]] =
     selectWhereList(fr"""userID = $userId AND (endedAt != NULL AND endedAt < NOW()::TIMESTAMP)""")
 
-  def futureUserCourses(userId: Long): TranzactIO[List[Course]] =
+  def futureUserCourses(userId: Int): TranzactIO[List[Course]] =
     selectWhereList(fr"""userID = $userId AND (startedAt != NULL AND  NOW()::TIMESTAMP < statedAt)""")
   
   def linkedToTemplateCourses(templateAlias: String): TranzactIO[Seq[Course]] =
