+++ zioServer/src/main/scala/tester/srv/dao/CourseTemplateForGroupDao.scala
@@ -13,16 +13,16 @@ import doobie.postgres.pgisimplicits.*
 
 
 object CourseTemplateForGroupDao extends AbstractDao[CourseTemplateForGroup] {
-  case class CourseTemplateForGroup(id: Long, groupId: Long, templateAlias: String, forceStartForGroupMembers: Boolean)
+  case class CourseTemplateForGroup(id: Int, groupId: Int, templateAlias: String, forceStartForGroupMembers: Boolean)
 
   override val schema: Schema[CourseTemplateForGroup] = DeriveSchema.gen[CourseTemplateForGroup]
   override val tableName: String = "CourseTemplateForGroup"
 
-  def removeCourseFromGroup(templateAlias: String, groupId: Long): TranzactIO[Int] =
+  def removeCourseFromGroup(templateAlias: String, groupId: Int): TranzactIO[Int] =
     deleteWhere(fr"templateAlias = $templateAlias AND groupId = $groupId")
 
   /** Курсы, которые должны стартовать для всех участников группы автоматически */
-  def forcedCourses(groupId: Long): TranzactIO[List[CourseTemplateForGroup]] =
+  def forcedCourses(groupId: Int): TranzactIO[List[CourseTemplateForGroup]] =
     selectWhereList(fr"groupId = $groupId")
 
 }
